+++ src/components/index.ts
@@ -1,3 +1,2 @@
 export { default as Banner } from "./Banner";
 export { default as Navbar } from "./Navbar";
-export { default as SearchBar } from "./SearchBar";
